wireless_manager: Handle HT40+/- correctly for 2.4GHz channels
Below is an excerpt from hostapd.conf:

	Note: There are limits on which channels can be used with HT40- and
	HT40+. Following table shows the channels that may be available for
	HT40- and HT40+ use per IEEE 802.11n Annex J:
	freq		HT40-		HT40+
	2.4 GHz		5-13		1-7 (1-9 in Europe/Japan)
	5 GHz		40,48,56,64	36,44,52,60

The current code only correctly handles 5 GHz channels.  When applied to
2.4GHz channels, channels 1,2, and 3 were being configured with HT40- which
caused hostapd to fail to start.

This change mimics the logic of openwrt's mac80211.sh.  For the 2.4GHz band,
channels less than 7 are set for HT40+ and everything else is set for HT40-.
This complies with the table above.

Fixes: 4f8d478a21cf ("calculate wireless capabilities dynamically")

NGFW-11553